% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nom_voiture}
\alias{nom_voiture}
\title{Noms des voitures}
\format{
Un data frame avec une seule colonne :
\describe{
\item{nom}{Le nom de la voiture.}
}
}
\source{
mtcars dataset
}
\usage{
nom_voiture
}
\description{
Ce jeu de données contient les noms des voitures à partir du jeu de données mtcars.
}
\examples{
head(nom_voiture)

}
\keyword{datasets}
