% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produit.R
\name{produit}
\alias{produit}
\title{Calculer le produit de deux nombres}
\usage{
produit(a, b)
}
\arguments{
\item{a}{Le premier nombre à multiplier}

\item{b}{Le deuxième nombre à multiplier}
}
\value{
Le produit de \code{a} et \code{b}
}
\description{
Cette fonction prend deux nombres en entrée et retourne leur produit.
}
\examples{
produit(2, 3) # Renvoie 6
produit(5, -4) # Renvoie -20

}
